fix(geometry): un-compact #4122's ratchet-fragile struct literals
#4122's patch (cherry-picked as the prior commit) got mesh.rs,
element.rs, router/voids/geom.rs and router/voids/mod.rs under the
module-size ratchet by collapsing struct-literal field lists onto one
line and trimming a doc comment, instead of shrinking the files. Every
one of those compactions is a rustfmt violation: `cargo fmt` reflows
them back to multi-line and pushes all four files back over budget
(1011->1012, 880->882, 2056->2058, 815->816 lines against those exact
budgets). None of that is exercised by CI (`cargo fmt --check` isn't
gated), so it stayed invisible until this review.

Restore normal multi-line formatting at all six sites the patch
compacted (three in mesh.rs's constructors, one in
voids/geom.rs's mesh_to_frame, one in voids/mod.rs's
translate_cutter_mesh, and element.rs's debug_assert!, which was one
136-column line), and recover the ratchet budget properly -- by
extracting to a sibling module (the house pattern, AGENTS.md) instead
of re-compacting:

- mesh.rs: Mesh::welded_by_position's ~120-line weld_impl helper moves
  to mesh_position_weld.rs (901 lines, was 1012 budget).
- voids/geom.rs: the frame-projection trio (wall_frame_from_depth,
  mesh_to_frame, project_aabb_in_frame) moves to frame_project.rs,
  re-exported so existing `geom::` call sites are unchanged (813
  lines, was 882 budget).
- voids/mod.rs: the malformed-cutter (#1007) repair cluster
  (OpeningBox + its box-mesh builder, the manifold check, the OBB
  detector, the re-cut pass, and the two frame-translation helpers)
  moves to malformed_opening_repair.rs, same re-export treatment
  (1739 lines, was 2058 budget).
- element.rs: build_mesh_data (the single MeshData-construction
  funnel, ~100 lines) moves to element_mesh_build.rs (716 lines, was
  816 budget).

Every extracted item keeps its original effective visibility
(`pub(in crate::router::voids)` where a plain `pub(super)` on the new
child module would have narrowed it below what sibling modules --
prism_cut.rs, synthesis/exit_cap.rs, flap_clip_tests.rs -- already
relied on). No production logic changed; this is a pure move plus
formatting. `rustfmt --check` on every touched and new file now shows
only baseline violations that predate #4122's patch and this repo's
untouched files already carry (confirmed against an unrelated,
unmodified file) -- none of the six sites this patch introduced.

welded_in_object_frame's debug_assert! itself is unchanged: #4122's
own "Shape of the fix" section specifies a debug_assert, this is
prerequisite instrumentation for the #4103 residual (welding at the
cached source), not a production behaviour change, and its silence in
release builds is intentional per that issue -- so this commit does
not touch build_mesh_data's actual `instance_meta.is_some()` branch.
